+++ src/asp/Core/StereoSettings.h
@@ -240,7 +240,6 @@ namespace asp {
 
     double min_triangulation_angle;           // min angle for valid triangulation
     double max_valid_triangulation_error;
-    bool   use_least_squares;                 // Use a more rigorous triangulation
     bool   save_double_precision_point_cloud; // Save final point cloud in double precision rather than bringing the points closer to origin and saving as float (marginally more precision at 2x the storage).
     double point_cloud_rounding_error;        // How much to round the output point cloud values
     bool   compute_point_cloud_center_only;   // Only compute the center of triangulated point cloud and exit.